# ResponseWriter and Message types