… docs/protocol.md
… go.sum
… pkg/server/mux_test.go