- [x] ResponseWriter
- [x] Request
- [x] Mux Updates
- [x] Marshaler / Unmarshaler
- [x] Use R/R
- [x] Append R/R
- [x] Query Req
- [x] Query Resp
- [x] Response Messages with Command
- [x] Stats

Closes #13 
Closes #24