@gideonw 2022-12-03
Able to achieve some perf improvements:

**Before**
```
> Test run finished at 12/3/2022, 2:54:02 AM <

Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkMapCommandParse$ github.com/dburkart/fossil/pkg/server

goos: darwin
goarch: arm64
pkg: github.com/dburkart/fossil/pkg/server
BenchmarkMapCommandParse
BenchmarkMapCommandParse-10     32620142                36.34 ns/op           48 B/op          1 allocs/op
PASS
ok      github.com/dburkart/fossil/pkg/server   2.914s

```
**After**
```
> Test run finished at 12/3/2022, 2:56:22 AM <

Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkMapCommandParse$ github.com/dburkart/fossil/pkg/server

goos: darwin
goarch: arm64
pkg: github.com/dburkart/fossil/pkg/server
BenchmarkMapCommandParse
BenchmarkMapCommandParse-10     136650238                8.520 ns/op           0 B/op          0 allocs/op
PASS
ok      github.com/dburkart/fossil/pkg/server   2.812s


> Test run finished at 12/3/2022, 3:00:32 AM <
```